Fix tabbar
Fixes https://github.com/lxde/qterminal/issues/395

The following issues are fixed:

(1) `Properties::Instance()->showCloseTabButton` shouldn't be used with `setTabsClosable()` in the c-tor of `TabWidget` because its value is not defined yet. This fixes the erratic behavior of tab close buttons.

(2) The tab pages shouldn't be deleted at `MainWindow::closeEvent()` -- except for one case -- because `TabWidget::removeTab()` deletes them; the last page is also deleted because it's a child of `TabWidget`.

(3) To make the slected tab text bold, we should give enough space to the tab by making the tabbar font bold and then, making the non-selected tab text normal.