Fix tab texts, especially for vertical tabs
Fixes https://github.com/lxqt/qterminal/issues/409

The custom style of the code was not only wrong but redundant. It was wrong -- like many other hard-coded styles -- because it ignored some aspects of `QStyle`. It was redundant because text eliding could be done in an easier way and style engines know how to draw icons, if any.

With this patch, not only vertical tabs are drawn correctly but also all tabs show the standard behavior: if the width limit is not set, tabs are shrinked to avoid tab scroll buttons as far as possible (previously, a single tab with a long text could make a useless scroll button show up).