Scrollable and compact Settings dialog
Closes https://github.com/lxqt/qterminal/issues/574

If the diff is seen inside a visual diff tool (like KDiff3), it'll be obvious that only a `QScrollArea` is added; the rest is just about indentation.

Also, it's tried to limit the dialog size to available desktop geometry. If the screen is too small, the user could resize the dialog by pressing `Alt` because the scroll area can be made smaller.