Adds virtual destructor to a base class
cppcheck report:
Class 'AbstractBookmarkItem' which is inherited by class
'BookmarkRootItem' does not have a virtual destructor. If you destroy
instances of the derived class by deleting a pointer that points to the
base class, only the destructor of the base class is executed. Thus,
dynamic memory that is managed by the derived class could leak.

Avoided by adding a virtual destructor to the base class.