Improved the Bookmarks page of QTerminal's settings dialog (#919)
* Improved the Bookmarks page of QTerminal's settings dialog

In short, this patch makes all widgets of the Bookmarks page work as expected. What follows is about details.

The default XML text is saved to a bookmarks file inside the user's config directory the first time he clicks the "Apply" or "OK" button of QTerminal's settings dialog.

An "Examples" button is added. It opens the app directory, which contains example bookmarks file(s). Also, if the path inside the line-edit is nonexistent, the same directory will be opened by clicking the "Find..." button; otherwise, the path will be selected in the file dialog.

If the user chooses an example bookmarks file from the app directory, it'll be copied to his config directory and will be chosen as soon as he confirms the settings dialog, so that he'll be able to edit it later.

Also, manual editing of the line-edit has effect (although most users may not use it):

 * If the user types a nonexistent file path inside the line-edit and confirms the settings dialog, the file will be created with the contents that are shown in the text view.
 * If the user types an existing, readable file path inside the line-edit, it'll be loaded when the line-edit loses focus — exactly as if the file was opened by pressing the "Find..." button.

* Made a string translatable

* Warn user if he wants to modify a non-xml bookmarks file