Fixed inactive tab text with Qt6
Due to a change (or bug?) in Qt6, using a stylesheet for setting the font of active/inactive tabs doesn't work as it did with Qt5.

Here, `QStyle` is used directly, which is always the best approach because stylesheets can interfere with the widget style in bad ways.

Fixes https://github.com/lxqt/qterminal/issues/1112