Prevent ordinary main windows under dropdown process
When a new main window is opened in the dropdown mode (menubar → File → New Window), it should have a separate process because the dropdown process is supposed to consist of a single dropdown window. Of course, the work directory and profile must be respected too.

Fixes https://github.com/lxqt/qterminal/issues/1153