Prompt on closing with a running process (#1200)
Previously, the closing prompt was useful only when there was a running process in the terminal, but the code didn't check that. A prompt without a running process was redundant.

This patch gives a real job to the closing prompt by making it work only when there is a running process, whether when the main window or a tab is going to be closed, or when a subterminal is going to be collapsed.

To detect a running process, the code makes use of @marcusbritanicus's commit https://github.com/lxqt/qtermwidget/commit/ccb3b2132d3c7d4b19b9a3221f9c4dcf13eee9dc by comparing `getForegroundProcessId()` with `getShellPID()`. In addition, it checks whether `TerminalConfig` has a command (e.g., when an app is executed in the terminal by the file manager).

Therefore, the config text "Ask for confirmation when closing" is also changed to "Prompt on closing with a running process".