Made modal dialogs stay in front of dropdown window on Wayland
That's done for any child dialog of the main window. It can't be done (consistently) for another dialog that may be called by a child dialog — especially, it's impossible for LXQt file dialog.

Closes https://github.com/lxqt/qterminal/issues/1133

NOTE: The main issue should be fixed by Wayland implementations. I made this patch only because I found a relatively simple and centralized workaround.