Removed "Do not ask again" from prompt dialog
Also, chose "No" as the default button.

The reason behind both changes is that, unlike in older versions of QTerminal, the prompt dialog has the important job of preventing closing when a process is running, and so, its disabling shouldn't be possible by mistake (see https://github.com/lxqt/qterminal/issues/274).

Closes https://github.com/lxqt/qterminal/issues/274